+++ docs/about/release-notes.rst
@@ -7,7 +7,7 @@ Release Notes
 .. Template, copy this to create a new section after a release:
 
    vRELEASE_PLACEHOLDER
-   --------------------
+   --------
 
    Features
    ~~~~~~~~
@@ -37,12 +37,12 @@ Release Notes
 
 
 vRELEASE_PLACEHOLDER
---------------------
+--------
 
 Features
 ~~~~~~~~
 
-* Added new string-formatting options `#3017 <https://github.com/scipp/scipp/pull/3017>`_.
+* Added new string-formatting options `#3017 <https://github.com/scipp/scipp/pull/3017>`_, `#3028 <https://github.com/scipp/scipp/pull/3028>`_.
 * Add ``size`` property to :class:`scipp.Variable` and :class:`scipp.DataArray`.
 
 Breaking changes